discoverer: migrate to go-config and go-storage
`go-config` provides better configuration inheritance and merging,
while `go-storage` offers a unified interface for etcd and Tarantool
storage backends. `tt/lib/cluster` is deprecated.

As part of the migration, normalize configuration key prefixes to
always include a leading slash to match `go-storage`'s `DefaultNamer`
behavior. Update all test keys accordingly.

+++ discoverer/discovery_test.go
@@ -472,9 +472,9 @@ groups:
 					Roles:      []string{"roles.tqe-orchestrator"},
 					RolesCfg: map[string]any{
 						"roles.tqe-orchestrator": map[string]any{
-							"partitions":        10,
-							"rebalance_timeout": 5,
-							"sync_timeout":      1,
+							"partitions":        int64(10),
+							"rebalance_timeout": float64(5),
+							"sync_timeout":      float64(1),
 						},
 					},
 				},
